# api/swagger.yaml
definitions:
...
  v4.Envelope:
    properties:
      archived:
        default: false
        description: Is the envelope archived?
        example: true
        type: boolean
      categoryId:
        description: ID of the category the envelope belongs to
        example: 878c831f-af99-4a71-b3ca-80deb7d793c1
        type: string
      createdAt:
        description: Time the resource was created
        example: "2022-04-02T19:28:44.491514Z"
        type: string
      deletedAt:
        description: Time the resource was marked as deleted
        example: "2022-04-22T21:01:05.058161Z"
        type: string
      id:
        description: UUID for the resource
        example: 65392deb-5e92-4268-b114-297faad6cdce
        type: string
      links:
        allOf:
        - $ref: '#/definitions/v4.EnvelopeLinks'
        description: Links to related resources
      name:
        description: Name of the envelope
        example: Groceries
        type: string
      note:
        description: Notes about the envelope
        example: For stuff bought at supermarkets and drugstores
        type: string
      updatedAt:
        description: Last time the resource was updated
        example: "2022-04-17T20:14:01.048145Z"
        type: string
    type: object
  v4.EnvelopeCreateResponse:
    properties:
      data:
        description: Data for the Envelope
        items:
          $ref: '#/definitions/v4.EnvelopeResponse'
        type: array
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
  v4.EnvelopeEditable:
    properties:
      archived:
        default: false
        description: Is the envelope archived?
        example: true
        type: boolean
      categoryId:
        description: ID of the category the envelope belongs to
        example: 878c831f-af99-4a71-b3ca-80deb7d793c1
        type: string
      name:
        description: Name of the envelope
        example: Groceries
        type: string
      note:
        description: Notes about the envelope
        example: For stuff bought at supermarkets and drugstores
        type: string
    type: object
  v4.EnvelopeLinks:
    properties:
      month:
        description: The MonthConfig for the envelope
        example: https://example.com/api/v4/envelopes/45b6b5b9-f746-4ae9-b77b-7688b91f8166/YYYY-MM
        type: string
      self:
        description: The envelope itself
        example: https://example.com/api/v4/envelopes/45b6b5b9-f746-4ae9-b77b-7688b91f8166
        type: string
      transactions:
        description: The envelope's transactions
        example: https://example.com/api/v4/transactions?envelope=45b6b5b9-f746-4ae9-b77b-7688b91f8166
        type: string
    type: object
  v4.EnvelopeListResponse:
    properties:
      data:
        description: List of Envelopes
        items:
          $ref: '#/definitions/v4.Envelope'
        type: array
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/v4.Pagination'
        description: Pagination information
    type: object
...
  v4.EnvelopeResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v4.Envelope'
        description: Data for the Envelope
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
...
  v4.MonthConfig:
    properties:
      allocation:
        description: The maximum value is "999999999999.99999999", swagger unfortunately
          rounds this.
        example: 22.01
        maximum: 1000000000000
        minimum: 1e-08
        multipleOf: 1e-08
        type: number
      envelopeID:
        description: We do not use the default model here, we use envelope ID and
          month
        type: string
      envelopeId:
        description: ID of the envelope
        example: 10b9705d-3356-459e-9d5a-28d42a6c4547
        type: string
      links:
        $ref: '#/definitions/v4.MonthConfigLinks'
      month:
        description: We do not use the default model here, we use envelope ID and
          month
        type: string
      note:
        description: A note for the month config
        example: Added 200€ here because we replaced Tim's expensive vase
        type: string
    type: object
  v4.MonthConfigEditable:
    properties:
      allocation:
        description: The maximum value is "999999999999.99999999", swagger unfortunately
          rounds this.
        example: 22.01
        maximum: 1000000000000
        minimum: 1e-08
        multipleOf: 1e-08
        type: number
      envelopeId:
        description: ID of the envelope
        example: 10b9705d-3356-459e-9d5a-28d42a6c4547
        type: string
      month:
        description: The month. This is always set to 00:00 UTC on the first of the
          month.
        example: "1969-06-01T00:00:00.000000Z"
        type: string
      note:
        description: A note for the month config
        example: Added 200€ here because we replaced Tim's expensive vase
        type: string
    type: object
  v4.MonthConfigLinks:
    properties:
      envelope:
        description: The Envelope this config belongs to
        example: https://example.com/api/v4/envelopes/61027ebb-ab75-4a49-9e23-a104ddd9ba6b
        type: string
      self:
        description: The Month Config itself
        example: https://example.com/api/v4/envelopes/61027ebb-ab75-4a49-9e23-a104ddd9ba6b/2017-10
        type: string
    type: object
  v4.MonthConfigResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v4.MonthConfig'
        description: Config for the month
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
...
  v4.Pagination:
    properties:
      count:
        description: The amount of records returned in this response
        example: 25
        type: integer
      limit:
        description: The maximum amount of resources to return for this request
        example: 25
        type: integer
      offset:
        description: The offset for the first record returned
        example: 50
        type: integer
      total:
        description: The total number of resources matching the query
        example: 827
        type: integer
    type: object
...
  v4.httpError:
    properties:
      error:
        example: An ID specified in the query string was not a valid UUID
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v4/envelopes:
    get:
      description: Returns a list of envelopes
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by note
        in: query
        name: note
        type: string
      - description: Filter by category ID
        in: query
        name: category
        type: string
      - description: Is the envelope archived?
        in: query
        name: archived
        type: boolean
      - description: Search for this text in name and note
        in: query
        name: search
        type: string
      - description: The offset of the first Envelope returned. Defaults to 0.
        in: query
        name: offset
        type: integer
      - description: Maximum number of Envelopes to return. Defaults to 50.
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.EnvelopeListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.EnvelopeListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.EnvelopeListResponse'
      summary: Get envelopes
      tags:
      - Envelopes
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      responses:
        "204":
          description: No Content
      summary: Allowed HTTP verbs
      tags:
      - Envelopes
    post:
      description: Creates a new envelope
      parameters:
      - description: Envelopes
        in: body
        name: envelope
        required: true
        schema:
          items:
            $ref: '#/definitions/v4.EnvelopeEditable'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v4.EnvelopeCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.EnvelopeCreateResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.EnvelopeCreateResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.EnvelopeCreateResponse'
      summary: Create envelope
      tags:
      - Envelopes
  /v4/envelopes/{id}:
    delete:
      description: Deletes an envelope
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Delete envelope
      tags:
      - Envelopes
    get:
      description: Returns a specific Envelope
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
      summary: Get Envelope
      tags:
      - Envelopes
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Allowed HTTP verbs
      tags:
      - Envelopes
    patch:
      consumes:
      - application/json
      description: Updates an existing envelope. Only values to be updated need to
        be specified.
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      - description: Envelope
        in: body
        name: envelope
        required: true
        schema:
          $ref: '#/definitions/v4.EnvelopeEditable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.EnvelopeResponse'
      summary: Update envelope
      tags:
      - Envelopes
  /v4/envelopes/{id}/{month}:
    get:
      description: Returns configuration for a specific month
      parameters:
      - description: ID of the Envelope
        in: path
        name: id
        required: true
        type: string
      - description: The month in YYYY-MM format
        in: path
        name: month
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
      summary: Get MonthConfig
      tags:
      - Envelopes
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      parameters:
      - description: ID of the Envelope
        in: path
        name: id
        required: true
        type: string
      - description: The month in YYYY-MM format
        in: path
        name: month
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Allowed HTTP verbs
      tags:
      - Envelopes
    patch:
      description: Changes configuration for a Month. If there is no configuration
        for the month yet, this endpoint transparently creates a configuration resource.
      parameters:
      - description: ID of the Envelope
        in: path
        name: id
        required: true
        type: string
      - description: The month in YYYY-MM format
        in: path
        name: month
        required: true
        type: string
      - description: MonthConfig
        in: body
        name: monthConfig
        required: true
        schema:
          $ref: '#/definitions/v4.MonthConfigEditable'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.MonthConfigResponse'
      summary: Update MonthConfig
      tags:
      - Envelopes
...
swagger: "2.0"